refactor(reactor-mcp): address post-review nits
- Remove express and @types/express dependencies (no longer imported)
- Rename src/express.ts to src/mcp-routes.ts to reflect actual content
- Fix duplicate comment on Sentry error handler in switchboard/server.ts
