+++ packages/reactor-mcp/package.json
@@ -43,12 +43,10 @@
     "ai": "5.0.0",
     "change-case": "catalog:",
     "document-model": "workspace:*",
-    "express": "^4.21.1",
     "vite": "catalog:",
     "zod": "catalog:"
   },
   "devDependencies": {
-    "@types/express": "^4.17.25",
     "@types/node": "catalog:",
     "vitest": "catalog:"
   }
