+++ packages/reactor-mcp/src/index.ts
@@ -1,4 +1,4 @@
-export { setupMcpServer, type SetupMcpServerOptions } from "./express.js";
+export { setupMcpServer, type SetupMcpServerOptions } from "./mcp-routes.js";
 export { createServer, type CreateServerOptions } from "./server.js";
 export * from "./tools/index.js";
 export * from "./tools/types.js";
