+++ packages/reactor-mcp/src/mcp-routes.ts
